KB-1-4

+++ src/main/kotlin/FirstLesson/FourthTask.kt
@@ -6,7 +6,7 @@ fun main() {
     val age: Byte = 27
     val partOfDay: Float = 0.075f
     val seconds: Short = 6480
-    val partOfYear: Double = 2.0547945205479453E-4
+    val partOfYear: Float = 2.0547945205479453E-4f
     val apogee: Int = 327000
 
     print("$length\n$age\n$partOfDay\n$seconds\n$partOfYear\n$apogee")
